SSM-6 - Fix the Standing and Sitting Tracker

+++ src/components/TimeTracker.tsx
@@ -1,28 +1,16 @@
 import { useEffect, useRef, useState } from "react";
 import { TbWalk } from "react-icons/tb";
-import { FaPlay } from "react-icons/fa6";
-import { MdSave } from "react-icons/md";
+import { MdSave, MdAirlineSeatReclineNormal } from "react-icons/md";
 import { GrPowerReset } from "react-icons/gr";
-
-/* 
-Start timer -> starts the sitCounter ✅
-Stop Sitting & Stand Now -> starts the standTimer & stops the sitCounter ✅
-Show Document title the active timer - ✅
-End Session - Resets sitCounter & standCounter, stores to the total Standing & Sitting time✅ 
-store to Local Storage -  ✅
-
-sittingDisabled = false, standingDisabled = true, endSessionDisabled = true
-sittingDisabled = true, standingDisabled = false, endSessionDisabled = true
-sittingDisabled = true, standingDisabled = true, endSessionDisabled = false
-*/
 
 export const TimeTracker = () => {
   const [sitCounter, setSitCounter] = useState<number>(0);
   const [standCounter, setStandCounter] = useState<number>(0);
   const [currentMode, setCurrentMode] = useState<"sit" | "stand" | null>(null);
-  const [sittingDisabled, setSittingDisabled] = useState<boolean>(false);
-  const [standingDisabled, setStandingDisabled] = useState<boolean>(true);
-  const [endSessionDisabled, setEndSessionDisabled] = useState<boolean>(true);
+  const [showLastSession, setShowLastSession] = useState<boolean>(false);
+  const [lastSittingTime, setLastSittingTime] = useState<string>("");
+  const [lastStandingTime, setLastStandingTime] = useState<string>("");
+  const [currentTime, setCurrentTime] = useState<string>("");
   const activeCounter = currentMode === "sit" ? sitCounter : standCounter;
 
   const sitIntervalRef = useRef<number | null>(null);
@@ -59,59 +47,80 @@ export const TimeTracker = () => {
   function startCounter(type: "sit" | "stand") {
     const ref = type === "sit" ? sitIntervalRef : standIntervalRef;
     const setFn = type === "sit" ? setSitCounter : setStandCounter;
-    if (ref.current) return;
     ref.current = setInterval(() => {
       setFn((prev) => prev + 1);
     }, 1000);
 
-    if (type === "sit") {
-      setSittingDisabled(true);
-      setStandingDisabled(false);
-    } else {
-      setStandingDisabled(true);
-      setEndSessionDisabled(false);
+    if (type === "sit" && standIntervalRef.current !== null) {
+      clearInterval(standIntervalRef.current);
+    }
+    // Fix: Remove unnecessary else block and directly check sitIntervalRef
+    if (type !== "sit" && sitIntervalRef.current !== null) {
+      clearInterval(sitIntervalRef.current);
     }
     setCurrentMode(type);
   }
-  function resetSitCounter() {
-    if (sitIntervalRef.current != undefined) {
-      clearInterval(sitIntervalRef.current);
-    }
-    const todayTotalSitting = parseInt(
-      localStorage.getItem("todayTotalSitting") ?? "0",
-      10
-    );
-    localStorage.setItem(
-      "todayTotalSitting",
-      (todayTotalSitting + sitCounter).toString()
-    );
-    setSitCounter(0);
-    startCounter("stand");
-    setCurrentMode("stand");
-    sitIntervalRef.current = null;
-  }
 
-  function endSession() {
-    if (standIntervalRef.current != undefined) {
+  function pauseSession() {
+    if (standIntervalRef.current !== null) {
       clearInterval(standIntervalRef.current);
     }
+
+    if (sitIntervalRef.current !== null) {
+      clearInterval(sitIntervalRef.current);
+    }
+  }
+
+  function saveAndResetData() {
     const todayTotalStanding = parseInt(
       localStorage.getItem("todayTotalStanding") ?? "0",
       10
+    );
+    const todayTotalSitting = parseInt(
+      localStorage.getItem("todayTotalSitting") ?? "0"
     );
     localStorage.setItem(
       "todayTotalStanding",
       (todayTotalStanding + standCounter).toString()
     );
+    localStorage.setItem(
+      "todayTotalSitting",
+      (todayTotalSitting + sitCounter).toString()
+    );
     setStandCounter(0);
+    setSitCounter(0);
     standIntervalRef.current = null;
-    setSittingDisabled(false);
-    setEndSessionDisabled(true);
+    sitIntervalRef.current = null;
+  }
+
+  function saveSession() {
+    const { hours, minutes, seconds } = formatTime(sitCounter);
+    const {
+      hours: hoursStand,
+      minutes: minutesStand,
+      seconds: secondsStand,
+    } = formatTime(standCounter);
+    setLastSittingTime(`${hours}h:${minutes}m:${seconds}s`);
+    setLastStandingTime(`${hoursStand}h:${minutesStand}m:${secondsStand}s`);
+
+    const now = new Date();
+    const formattedTime = now.toLocaleTimeString([], {
+      hour: "2-digit",
+      minute: "2-digit",
+    });
+    setCurrentTime(formattedTime);
+
+    setShowLastSession(true);
+
+    saveAndResetData();
   }
 
   function resetAllDataLocalStorage() {
-    localStorage.setItem("todayTotalStanding", "0");
-    localStorage.setItem("todayTotalSitting", "0");
+    if (window.confirm("Are you sure?")) {
+      localStorage.setItem("todayTotalStanding", "0");
+      localStorage.setItem("todayTotalSitting", "0");
+      setShowLastSession(false);
+    }
   }
 
   const sitTime = formatTime(sitCounter);
@@ -127,14 +136,13 @@ export const TimeTracker = () => {
           </div>
           <button
             className={`flex items-center justify-center gap-2  w-full ${
-              sittingDisabled
-                ? "disabled:bg-gray-400 disabled:cursor-not-allowed"
-                : "!bg-green-500 text-white"
+              activeCounter === sitCounter
+                ? "bg-green-500 text-white"
+                : "bg-gray-200"
             }`}
             onClick={() => startCounter("sit")}
-            disabled={sittingDisabled}
           >
-            <FaPlay /> Start timer
+            <MdAirlineSeatReclineNormal /> Track Sitting
           </button>
         </div>
 
@@ -145,17 +153,41 @@ export const TimeTracker = () => {
           </div>
           <button
             className={`flex items-center justify-center gap-2 ${
-              standingDisabled
-                ? "disabled:bg-gray-400 disabled:cursor-not-allowed"
-                : "bg-green-500 text-white"
+              activeCounter === standCounter
+                ? "bg-green-500 text-white"
+                : "bg-gray-200"
             }`}
-            onClick={resetSitCounter}
-            disabled={standingDisabled}
+            onClick={() => startCounter("stand")}
           >
-            <TbWalk /> Stand Now
+            <TbWalk /> Track Standing
           </button>
         </div>
       </div>
+      {showLastSession && (
+        <div className="bg-gray-50 border border-gray-200  rounded-lg mx-4 mb-4 py-4">
+          <div className="flex justify-around">
+            <div className="flex justify-center items-center gap-2">
+              <MdAirlineSeatReclineNormal className="text-xl" />
+              <div>
+                <p className="text-sm text-gray-500 ">Last Sitting Time</p>
+                <p className="font-bold text-gray-800">{lastSittingTime}</p>
+              </div>
+            </div>
+            <div className="flex justify-center items-center gap-2">
+              <TbWalk className="text-xl" />
+              <div>
+                <p className="text-sm text-gray-500">Last Stand Time</p>
+                <p className="font-bold text-gray-800">{lastStandingTime}</p>
+              </div>
+            </div>
+          </div>
+
+          <p className="text-xs text-gray-400 pl-8 mt-2 flex justify-items-start">
+            Saved at {currentTime}
+          </p>
+        </div>
+      )}
+
       <div className="flex flex-col bg-purple-200 py-4 mx-4 rounded-lg">
         <p className="text-sm font-bold">Total 🧘 Today</p>
         <p className="text-2xl font-bold">{` ${formattedTodayTotalSitting.hours}h: ${formattedTodayTotalSitting.minutes}m: ${formattedTodayTotalSitting.seconds}s`}</p>
@@ -167,12 +199,16 @@ export const TimeTracker = () => {
       <div className="flex flex-col">
         <button
           className="flex items-center justify-center gap-2 bg-yellow-500 w-48 mx-auto text-white"
-          onClick={endSession}
+          onClick={pauseSession}
+          onDoubleClick={saveSession}
         >
           <MdSave /> Pause/Save
         </button>
+        <p className="mt-2 text-xs text-gray-400">
+          Double Click to Save Session
+        </p>
         <button
-          className="flex items-center justify-center gap-2 bg-red-400 mt-4 shadow-md w-48 mx-auto text-white"
+          className="flex items-center justify-center gap-2 bg-red-400 mt-2 shadow-md w-48 mx-auto text-white"
           onClick={resetAllDataLocalStorage}
         >
           <GrPowerReset /> Reset for Today
